requirements.txt
certifi~=2022.6.15
charset-normalizer~=2.1.0
humanize~=3.14.0
idna~=3.3
python-qbittorrent~=0.4.2
# requests==2.27.1
setuptools==59.6.0
urllib3~=1.26.10
wheel~=0.37.1
pytz~=2022.1
aiohttp~=3.8.1
pip==21.3.1
asyncio~=3.4.3